+++ src/main/java/net/thenextlvl/gopaint/brush/setting/CraftItemBrushSettings.java
@@ -64,7 +64,7 @@ public final class CraftItemBrushSettings implements ItemBrushSettings {
         return random;
     }
 
-    @Deprecated(forRemoval = true)
+    @Deprecated(forRemoval = true, since = "1.1.1")
     public static ItemBrushSettings parse(Brush brush, ItemMeta itemMeta) {
         var builder = builder().brush(brush).surfaceMode(SurfaceMode.DISABLED);
         Optional.ofNullable(itemMeta.getLore()).ifPresent(lore -> lore.stream()
