+++ src/main/java/net/thenextlvl/gopaint/command/GoPaintCommand.java
@@ -38,7 +38,7 @@ public class GoPaintCommand {
 
     private int size(CommandContext<CommandSourceStack> context) {
         var player = (Player) context.getSource().getSender();
-        var brush = plugin.brushController().getBrush(player);
+        var brush = plugin.brushController().getBrushSettings(player);
         brush.setSize(context.getArgument("size", int.class));
         plugin.bundle().sendMessage(player, "command.gopaint.brush.size",
                 Placeholder.parsed("size", String.valueOf(brush.getSize())));
@@ -47,7 +47,7 @@ public class GoPaintCommand {
 
     private int toggle(CommandContext<CommandSourceStack> context) {
         var player = (Player) context.getSource().getSender();
-        var brush = plugin.brushController().getBrush(player);
+        var brush = plugin.brushController().getBrushSettings(player);
         var message = brush.toggle() ? "command.gopaint.brush.enabled"
                 : "command.gopaint.brush.disabled";
         plugin.bundle().sendMessage(player, message);
